fix: Don't trust the length reported by the ExactSizeIterator trait
Since the `ExactSizeIterator` trait is not an unsafe trait we can't trust the length it reports when creating the slice. Doing so would make it possible to pass an iterator to `alloc_extend` which is larger than the length it reports, causing the arena to be reallocated.

As a bonus `alloc_extend` now only requires the `I: IntoIterator<Item = T>` trait