Support no_std by gating std usage on a "std" feature flag
This adds an enabled-by-default "std" cargo feature which can be selectively
removed to enable no_std support.

When the "std" feature is absent, feature(alloc) will be enabled to
provide Vec support. This requires nightly, but since "std" is enabled by
default, this should have no impact on stable users.